fix interpreter

+++ compiler/luci-interpreter/src/kernels/PadV2.cpp
@@ -100,7 +100,7 @@ void PadV2::execute() const
     case DataType::S16:
     {
       assert(output()->zero_point() == 0);
-      const auto pad_value = getTensorData<uint8_t>(constant_values())[0];
+      const auto pad_value = getTensorData<int16_t>(constant_values())[0];
       tflite::reference_ops::Pad(params, getTensorShape(input()), getTensorData<int16_t>(input()),
                                  &pad_value, getTensorShape(output()),
                                  getTensorData<int16_t>(output()));
